# docs/apiDocs.yml
openapi: 3.0.1
info:
  title: inTiVi API docs
  description: "inTiVi docs for your pleasure. Please use 'Authorize' with valid JWT to access protected endpoints."
  version: 1.0.0
...
paths:
...
  /api/mylist/{movieId}:
    put:
      tags:
        - my list
      summary: Add movie to user's movie list.
      description: Add movie to user's movie list.
      parameters:
        - in: path
          name: movieId
          required: true
          description: Id of a movie to add to user's list of movies.
          schema:
            type: integer
      security: []
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                type: array
        400:
          description: "error - bad request: missing required fields or invalid values."
          content:
            application/json:
              schema:
                type: array

    delete:
      tags:
        - my list
      summary: Delete movie from user's movie list.
      description: Delete movie from user's movie list.
      parameters:
        - in: path
          name: movieId
          required: true
          description: Id of a movie to delete from user's list of movies.
          schema:
            type: integer
      security: []
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                type: array
        400:
          description: "error - bad request: missing required fields or invalid values."
          content:
            application/json:
              schema:
                type: array
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []